requirements.txt
Django==2.1.5
django-auth-kerberos==1.2.5
django-auth-ldap==1.7.0
django-cors-headers==2.4.0
djangorestframework==3.9.0
kerberos==1.3.0
mysqlclient==1.3.14
pyasn1==0.4.5
pyasn1-modules==0.2.3
pykerberos==1.2.1
python-ldap==3.1.0
pytz==2018.9
